# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIHSafe Manizales API - HIV risk assessment, AI chatbot, and education
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface QuestionOption {
  value: string;
  label: string;
}

export type QuestionType = (typeof QuestionType)[keyof typeof QuestionType];

export const QuestionType = {
  single: "single",
  multi: "multi",
  number: "number",
  country: "country",
} as const;

export type QuestionDomain =
  (typeof QuestionDomain)[keyof typeof QuestionDomain];

export const QuestionDomain = {
  demographic: "demographic",
  behavioral: "behavioral",
  epidemiological: "epidemiological",
  clinical: "clinical",
} as const;

export type QuestionShowIf = {
  questionId?: string;
  equals?: string[];
};

export interface Question {
  id: string;
  order: number;
  text: string;
  helpText?: string;
  type: QuestionType;
  domain: QuestionDomain;
  conditional: boolean;
  showIf?: QuestionShowIf;
  options?: QuestionOption[];
  defaultValue?: string;
}

export type AnswerValue = string | number | string[];

export interface AssessmentAnswer {
  questionId: string;
  value: AnswerValue;
}

export interface AssessmentSubmission {
  answers: AssessmentAnswer[];
}

export interface DomainScore {
  domain: string;
  score: number;
  maxScore: number;
}

export type RiskFactorSeverity =
  (typeof RiskFactorSeverity)[keyof typeof RiskFactorSeverity];

export const RiskFactorSeverity = {
  low: "low",
  moderate: "moderate",
  high: "high",
} as const;

export interface RiskFactor {
  label: string;
  severity: RiskFactorSeverity;
}

export type AssessmentResultRiskLevel =
  (typeof AssessmentResultRiskLevel)[keyof typeof AssessmentResultRiskLevel];

export const AssessmentResultRiskLevel = {
  low: "low",
  moderate: "moderate",
  high: "high",
  very_high: "very_high",
} as const;

export interface AssessmentResult {
  id: string;
  riskLevel: AssessmentResultRiskLevel;
  riskScore: number;
  summary: string;
  recommendations: string[];
  factors: RiskFactor[];
  domainScores: DomainScore[];
  createdAt: string;
}

export interface RecentAssessment {
  id: string;
  riskLevel: string;
  ageRange: string;
  createdAt: string;
}

export type StatsSummaryRiskDistributionItem = {
  level: string;
  count: number;
};

export interface StatsSummary {
  totalAssessments: number;
  riskDistribution: StatsSummaryRiskDistributionItem[];
  modulesCompleted: number;
  chatMessages: number;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatRequest {
  messages: ChatMessage[];
}

export interface ChatResponse {
  reply: string;
  suggestions?: string[];
}

export type EducationModuleCategory =
  (typeof EducationModuleCategory)[keyof typeof EducationModuleCategory];

export const EducationModuleCategory = {
  prevention: "prevention",
  transmission: "transmission",
  testing: "testing",
  treatment: "treatment",
  stigma: "stigma",
} as const;

export interface EducationModule {
  id: string;
  title: string;
  description: string;
  durationMinutes: number;
  category: EducationModuleCategory;
  icon: string;
}

export interface EducationSection {
  heading: string;
  body: string;
}

export interface EducationFact {
  statement: string;
  isMyth: boolean;
  explanation?: string;
}

export interface QuizQuestion {
  id: string;
  prompt: string;
  options: string[];
  correctOption: number;
  explanation?: string;
}

export interface EducationModuleDetail {
  id: string;
  title: string;
  description: string;
  sections: EducationSection[];
  facts: EducationFact[];
  quiz: QuizQuestion[];
}

export interface QuizAnswer {
  questionId: string;
  selectedOption: number;
}

export interface QuizSubmission {
  moduleId: string;
  answers: QuizAnswer[];
}

export type QuizResultFeedbackItem = {
  questionId: string;
  correct: boolean;
  explanation?: string;
};

export interface QuizResult {
  score: number;
  total: number;
  passed: boolean;
  feedback: QuizResultFeedbackItem[];
}

export interface Clinic {
  id: string;
  name: string;
  address: string;
  phone: string;
  services: string[];
  hours: string;
  free?: boolean;
}

export type GetQuestions200 = {
  questions: Question[];
};

export type GetRecentAssessments200 = {
  items: RecentAssessment[];
};

export type GetEducationModules200 = {
  modules: EducationModule[];
};

export type GetClinics200 = {
  clinics: Clinic[];
};