# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VIHSafe Manizales API - HIV risk assessment, AI chatbot, and education
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chatbot/message:
    post:
      operationId: sendChatMessage
      tags: [chatbot]
      summary: Send a message to the AI educational chatbot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Chatbot reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
...
  /education/modules/{moduleId}:
    get:
      operationId: getEducationModule
      tags: [education]
      summary: Get a single educational module with full content
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Module detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EducationModuleDetail"
  /education/quiz/submit:
    post:
      operationId: submitQuiz
      tags: [education]
      summary: Submit answers for an educational quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
...
components:
  schemas:
...
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
    ChatRequest:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
    ChatResponse:
      type: object
      required: [reply]
      properties:
        reply:
          type: string
        suggestions:
          type: array
          items:
            type: string
...
    EducationSection:
      type: object
      required: [heading, body]
      properties:
        heading:
          type: string
        body:
          type: string
    EducationFact:
      type: object
      required: [statement, isMyth]
      properties:
        statement:
          type: string
        isMyth:
          type: boolean
        explanation:
          type: string
    QuizQuestion:
      type: object
      required: [id, prompt, options, correctOption]
      properties:
        id:
          type: string
        prompt:
          type: string
        options:
          type: array
          items:
            type: string
        correctOption:
          type: integer
        explanation:
          type: string
    EducationModuleDetail:
      type: object
      required: [id, title, description, sections, facts, quiz]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/EducationSection"
        facts:
          type: array
          items:
            $ref: "#/components/schemas/EducationFact"
        quiz:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
    QuizAnswer:
      type: object
      required: [questionId, selectedOption]
      properties:
        questionId:
          type: string
        selectedOption:
          type: integer
    QuizSubmission:
      type: object
      required: [moduleId, answers]
      properties:
        moduleId:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswer"
    QuizResult:
      type: object
      required: [score, total, passed, feedback]
      properties:
        score:
          type: integer
        total:
          type: integer
        passed:
          type: boolean
        feedback:
          type: array
          items:
            type: object
            required: [questionId, correct]
            properties:
              questionId:
                type: string
              correct:
                type: boolean
              explanation:
                type: string
...