# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIHSafe Manizales API - HIV risk assessment, AI chatbot, and education
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";

export interface ChatRequest {
  messages: ChatMessage[];
}