# lib/api-zod/src/generated/types/educationModuleDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIHSafe Manizales API - HIV risk assessment, AI chatbot, and education
 * OpenAPI spec version: 0.1.0
 */
import type { EducationFact } from "./educationFact";
import type { EducationSection } from "./educationSection";
import type { QuizQuestion } from "./quizQuestion";

export interface EducationModuleDetail {
  id: string;
  title: string;
  description: string;
  sections: EducationSection[];
  facts: EducationFact[];
  quiz: QuizQuestion[];
}